remove "explored" node property
It turns out that the "visited" property governs what
RoutePlanner::Node::FindNeighbors can actually find.
This means a portion of the A* algorithm, not searching through
already explored states, is already implemented for us.
However it's not implemented correctly.
States that have been visited once should still be considered as neighbors
until they have become "explored", i.e. until we are sure we  already
found the shortest path to them.

Evidence of the problems is that setting the heuristic to 0,
which should reduce to uniform cost search, yields bogus results.